requirements.txt
Flask
Flask-SQLAlchemy
bcrypt
python-dotenv
Werkzeug
boto3